+++ firmware/config/engines/honda_k_dbc.cpp
@@ -6,6 +6,7 @@
  */
 
 #include "honda_k_dbc.h"
+#include "table_helper.h"
 
 #if HW_PROTEUS & EFI_PROD_CODE
 #include "proteus_meta.h"
@@ -42,8 +43,17 @@ void setHondaK() {
 	engineConfiguration->injectionMode = IM_SEQUENTIAL;
 
 	engineConfiguration->crankingIACposition = 70;
-    engineConfiguration->postCrankingFactor = 1.25;
-    engineConfiguration->postCrankingDurationSec = 15;
+    
+	static const float defaultPostCrankingCLTBins[] = {
+		-20.0f, 0.0f, 30.0f, 60.0f
+	};
+	static const uint16_t defaultPostCrankingDurationBins[] = {
+		0, 2, 4, 6, 8, 10, 12, 15
+	};
+	copyArray(engineConfiguration->postCrankingCLTBins, defaultPostCrankingCLTBins);
+	copyArray(engineConfiguration->postCrankingDurationBins, defaultPostCrankingDurationBins);
+	setTable(engineConfiguration->postCrankingFactor, 1.0f);
+
     engineConfiguration->useRunningMathForCranking = true;
 
 	strcpy(engineConfiguration->engineMake, ENGINE_MAKE_HONDA);