+++ firmware/CHANGELOG.md
@@ -30,6 +30,7 @@ Release template (copy/paste this for new release):
 
 ### Breaking Changes
  - "acIdleRpmBump" renamed to "acIdleRpmTarget", and changed the way of RPM rise needed for proper A/C operation from added to absolute target #5628
+ - After Cranking Enrichment changed from fixed value to table, it can help with some engines that want lots of fuel on cold, but don't run too well with big enrichment on hot
 
 ### Added
  - DAC with Lua #5601